Добавлены остальные страницы, вынесена страница с работниками
