+++ SergeiLevin0/Views/Employees/EmployeeDatails.cshtml
@@ -12,7 +12,7 @@
 <body>
 <h3>@ViewBag.Title</h3>
 <h4>@ViewData["Description"]</h4>
-<table>
+<table class="table">
     <thead>
     <tr>
         <th>Id</th>
@@ -34,6 +34,6 @@
     </tr>
     </tbody>
 </table>
-<p><a asp-controller="Home" asp-action="GetEmployees">Return to list of empoyees!</a></p>
+<p><a asp-controller="Employees" asp-action="Index">Return to list of empoyees!</a></p>
 </body>
 </html>
