+++ SergeiLevin0/Views/Shared/Partial/_Styles.cshtml
@@ -1,0 +1,7 @@
+﻿<link href="~/css/bootstrap.min.css" rel="stylesheet">
+<link href="~/css/font-awesome.min.css" rel="stylesheet">
+<link href="~/css/prettyPhoto.css" rel="stylesheet">
+<link href="~/css/price-range.css" rel="stylesheet">
+<link href="~/css/animate.css" rel="stylesheet">
+<link href="~/css/main.css" rel="stylesheet">
+<link href="~/css/responsive.css" rel="stylesheet">
